\name{parameters}
\alias{parameters}
\alias{T}
\alias{W}
\alias{M}
\alias{A}
\alias{bins}
\title{Logicle Scale Parameters}
\description{
	Returns the standard parameter values associated with the opaque parameters object.
}

\usage{
	T(p)
	W(p)
	M(p)
	A(p)
	bins(p)
}
\arguments{
	\item{p}{The Logicle parameters to be interrogated}
}

\references{ Logicle Manuscript. 2007. Logicle Tech Note. 2010 }
\author{Wayne A. Moore, Genetics Department, Stanford University}
\keyword{logicle}
\keyword{biexponential}
\keyword{vector}
